+++ packages/request-generate/request.config.ts
@@ -2,8 +2,12 @@
  * @type {import('@gopowerteam/request-generate').GenerateOptions}
  */
 module.exports = {
-  gateway: 'https://mall-service.gopowerteam.cn/admin',
-  openapi: '/api-docs',
+  gateway: 'https://gateway.local.xbt.sx.cn',
+  applications: {
+    'mall-service': 'xbt-platform-mall-service',
+    'file-service': 'xbt-platform-file-service',
+  },
+  openapi: '/v2/api-docs',
   output: './generated',
   exportModels: true,
   logger: true,