+++ packages/request-generate/src/generate/write-models.ts
@@ -2,7 +2,6 @@ import type { GenerateClient } from '../types/generate-client'
 import type { GenerateApplicationOptions } from '../types/generate-options'
 import * as fs from 'node:fs'
 import * as path from 'node:path'
-import { updateProgress } from '../progress'
 import { writeModel } from './write-model'
 
 /**
@@ -33,6 +32,5 @@ export function writeModels(
   client.models.forEach((model) => {
     const filename = `${model.name}.ts`
     writeModel(model, path.join(output, filename))
-    updateProgress(options.name || 'default', 'model')
   })
 }